fixed background service to be sticky and not require ongoing notification
