+++ core/src/main/java/org/solovyev/android/messenger/App.java
@@ -195,7 +195,7 @@ public final class App implements SharedPreferences.OnSharedPreferenceChangeList
 			final Boolean startOnBoot = MessengerPreferences.startOnBoot.getPreference(preferences);
 			enableComponent(application, OnBootBroadcastReceiver.class, startOnBoot);
 		} else if (MessengerPreferences.isOngoingNotificationEnabled.isSameKey(key)) {
-            refreshBackgroundService();
+            startBackgroundService();
         }
 	}
 	/*
@@ -308,7 +308,7 @@ public final class App implements SharedPreferences.OnSharedPreferenceChangeList
 		getAccountConnectionsService().tryStopAll();
 
 		final Intent serviceIntent = new Intent();
-		serviceIntent.setClass(instance.application, OngoingNotificationService.class);
+		serviceIntent.setClass(instance.application, StickyService.class);
 		instance.application.stopService(serviceIntent);
 
 		activity.finish();
@@ -348,21 +348,10 @@ public final class App implements SharedPreferences.OnSharedPreferenceChangeList
 		return MessengerPreferences.Gui.theme.getPreferenceNoError(getPreferences());
 	}
 
-	public static void refreshBackgroundService() {
-        final SharedPreferences preferences = getPreferences();
-        final boolean isOngoingNotificationEnabled = MessengerPreferences.isOngoingNotificationEnabled.getPreference(preferences);
-
+	public static void startBackgroundService() {
         final Intent serviceIntent = new Intent();
-        serviceIntent.setClass(instance.application, OngoingNotificationService.class);
-
-        if (isOngoingNotificationEnabled && !OngoingNotificationService.isRunning)
-        {
-            instance.application.startService(serviceIntent);
-        }
-        else if (!isOngoingNotificationEnabled && OngoingNotificationService.isRunning)
-        {
-            instance.application.stopService(serviceIntent);
-        }
+        serviceIntent.setClass(instance.application, StickyService.class);
+        instance.application.startService(serviceIntent);
     }
 
 	public static boolean isDebuggable() {
