+++ core/src/main/java/org/solovyev/android/messenger/OnBootBroadcastReceiver.java
@@ -20,12 +20,12 @@ import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 
-import static org.solovyev.android.messenger.App.refreshBackgroundService;
+import static org.solovyev.android.messenger.App.startBackgroundService;
 
 public class OnBootBroadcastReceiver extends BroadcastReceiver {
 
 	@Override
 	public void onReceive(Context context, Intent intent) {
-		refreshBackgroundService();
+		startBackgroundService();
 	}
 }
